+++ src/getMonthDays.ts
@@ -1,10 +1,15 @@
 import { NaiveDate } from "./NaiveDate";
 import { nextMonth } from "./YearMonth";
 
+export enum StartOfWeek {
+  Sunday = "sunday",
+  Monday = "monday",
+}
+
 export function getMonthDays(
   year: number,
   month: number,
-  startOfWeek: "S" | "M" = "M",
+  startOfWeek: StartOfWeek,
 ): [number[], number] {
   const date = new NaiveDate(year, month).toLocalDate();
   const dayCount = (() => {
@@ -17,6 +22,8 @@ export function getMonthDays(
   })();
   return [
     Array.from(Array(Math.round(dayCount)), (_, i) => i + 1),
-    startOfWeek === "S" ? date.getDay() : (date.getDay() + 6) % 7,
+    startOfWeek === StartOfWeek.Sunday
+      ? date.getDay()
+      : (date.getDay() + 6) % 7,
   ];
 }