+++ src/index.ts
@@ -1,5 +1,6 @@
+export * from "./colors";
+export { DatePicker, DatePickerProps } from "./DatePicker";
 export { NaiveDate } from "./NaiveDate";
 export { NaiveTime } from "./NaiveTime";
-export { DatePicker } from "./DatePicker";
-export { TimePicker } from "./TimePicker";
-export * from "./colors";
+export { StartOfWeek } from "./getMonthDays";
+export { TimePicker, TimePickerProps } from "./TimePicker";